% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_growth.R
\name{detailed_growth}
\alias{detailed_growth}
\title{Calculate between-census annual growth}
\usage{
detailed_growth(data)
}
\arguments{
\item{data}{A data frame containing repeated dbh measurements of trees. Must 
have a column named "tree_id" that contains unique tree ID values, a column
named "year" containing the year of the measurement, and a column named
"dbh" containing the dbh measurements. See built-in dataset \code{tree} for
an example.}
}
\value{
A data frame containing between-census annual growth of each tree:
\describe{
  \item{tree_id}{unique identification code of the tree}
  \item{start_year}{first year of between-census period}
  \item{begin_size}{diameter at breast height at beginning of between-census
  period, in cm}
  \item{end_year}{last year of between-census period}
  \item{end_size}{diameter at breast height at end of between-census
  period, in cm}
  \item{annual_growth}{annual growth rate as average yearly increase in 
  diameter at breast height, in cm/year}
  \item{annual_bai}{annual basal area increment as average yearly increase
  in area at breast height, in cm^2/year}
  \item{size_corr_growth}{square root of \code{annual_growth} divided by
  \code{begin_size} to give somewhat normally distributed growth rates}
  \item{...}{any other columns that appeared in \code{data}}
}
}
\description{
Requires a data frame where separate measurements of the same tree appear
in different rows and returns a data frame containing annual growth between
each pair of consecutive measurements for each tree.
}
\examples{
detailed_growth(tree)
}
