% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_check.R
\name{mapping_check}
\alias{mapping_check}
\title{Check tree mapping data}
\usage{
mapping_check(map_data, max_x, max_y)
}
\arguments{
\item{map_data}{Data frame containing tree mapping data. Should contain the
columns \code{tree_id}, \code{stand_id}, \code{species}, \code{x_coord},
and \code{y_coord}. Any additional columns will be ignored by this function.}

\item{max_x}{Maximum expected x coordinate (i.e. should be 100 if the stands
are 100 x 100 m).}

\item{max_y}{Maximum expected y coordinate.}
}
\value{
A list containing two elements:
\itemize{
  \item{\code{problem_trees} is a data frame containing the rows of
  \code{map_data} that contain data issues. An additional column describes
  the identified issue}
  \item{\code{issue_summary} is a data frame that shows the number and 
  percentage of trees with at least one issue and with each of the specific
  issues}
}
}
\description{
Checks a tree mapping dataset for missing data and formatting that could 
lead to errors when applying other functions in this package. Outputs a
table of the trees with data issues and a table summarizing the number of
trees in the mapping dataset that have data issues.
}
\details{
The data issues checked for are: presence of required columns, duplicated
tree ids, missing x or y coordinates, x or y coordinates outside the
expected range, missing stand id or species information. This function does
not check for misspelled stand ids or species, which should be checked 
independently.
}
\examples{
map_check_test <- mapping_check(messy_mapping, 100, 100)
}
