% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_summary.R
\name{neighborhood_summary}
\alias{neighborhood_summary}
\title{Summarize neighborhoods}
\usage{
neighborhood_summary(neighbors, radius, densities = "raw")
}
\arguments{
\item{neighbors}{Data frame of neighborhoods outputted by
\code{neighborhoods} function.}

\item{radius}{Numeric vector describing neighborhood radius in meters.}

\item{densities}{Character specifying the type of density measurements to
calculate - raw (m^2 per hectare), proportional (as proportions of overall
tree density), angular (angular size of trees).}
}
\value{
Data frame containing a summary for each neighborhood in
\code{neighbors}.
}
\description{
Calculates species-specific densities and species richness for each 
neighborhood in the provided data frame. Densities can be provided as
area covered in the neighborhood, area covered per hectare, or the sum of
horizontal angles.
}
