% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_nbhd_size.R
\name{select_nbhd_size}
\alias{select_nbhd_size}
\title{Select neighborhood size}
\usage{
select_nbhd_size(
  radii,
  map_data,
  growth_data,
  abiotic_data = NULL,
  focal_sps,
  dens_type,
  edge_handling = F,
  max_x,
  max_y,
  rare_sps = 100
)
}
\arguments{
\item{radii}{Vector of the neighborhood radii to try.}

\item{map_data}{Data frame containing tree coordinates.}

\item{growth_data}{Data frame containing repeated measurements of DBH for
trees in \code{map_data}}

\item{abiotic_data}{Data frame containing abiotic data for each of the stands
in \code{map_data}. Must contain a stand_id column. All other columns will be
included as explanatory variables in the fitted models. This argument is
optional.}

\item{focal_sps}{Vector of species names (i.e. character strings) for which
a neighborhood size should be selected.}

\item{dens_type}{Method for calculating tree densities in each neighborhood.
Must take one of the values: \code{"raw"}, \code{"proportional"}, or
\code{"angular"}. See \code{?neighborhood_summary} for details.}

\item{edge_handling}{Boolean indicating whether trees whose neighborhood
overlaps the stand boundary should have their neighborhood metrics adjusted
to account for the unsampled portion of their neighborhood. If FALSE, these
trees will be excluded from the analysis as focals.}

\item{max_x}{Maximum expected x coordinate (i.e. should be 100 if the stands
are 100 x 100 m).}

\item{max_y}{Maximum expected y coordinate.}

\item{rare_sps}{Minimum number of interactions a competitor species must
appear in to remain separate from the "RARE" category in the growth models
(see \code{?growth_model} for details). If not specified, this argument will
default to a value of 100 interactions.}
}
\value{
A list containing a number of named elements (exact number will be
the number of \code{focal_sps} plus one:
\itemize{
\item \code{mse_vals} is a data frame containing the mean square error of
the fitted model for each \code{focal_sps} using each neighborhood size
listed in \code{radii}.
\item each additional element of the list will be a plot of model mean square
error vs. neighborhood size for one of the \code{focal_sps}.
}
}
\description{
Fits tree growth models using neighborhood metrics based on differing
neighborhood sizes, calculating mean square error of each model, thereby
providing useful data on the appropriate neighborhood size for analysis. It
is strongly recommended that this sensitivity analysis or a similar one is
used to select a neighborhood size for modeling because the appropriate 
neighborhood size varies between study systems and ecological processes.
}
\details{
This function uses a data frame of tree coordinates (\code{map_data}) to
calculate a series of metrics to describe the neighborhoods around each tree.
It does this by applying the \code{neighborhoods} and then the
\code{neighborhood_summary} function. It is therefore necessary to provide,
under the argument \code{dens_type}, instruction for the method that
\code{neighborhood_summary} should use to calculate densities. It is also
possible to conduct edge handling by setting \code{edge_handling = T}. This
process is repeated for each of the neighborhood sizes specified by the
argument \code{radii}. To prevent edge effects, trees whose neighborhood
overlaps the stand boundary are excluded from modeling (hence the need to
provide stand sizes with the arguments \code{max_x} and \code{max_y}).
Boundary overlap is determined using the largest neighborhood size in
\code{radii} to ensure the number of focal trees is constant across
neighborhood sizes.

The neighborhood metrics are then combined with the growth rates of their
focal tree, specifically the \code{size_corr_growth} output by 
\code{growth_summary} (any trees with illogical negative annual growth rates
are excluded as focal trees). If the user also provides abiotic data for the
stands using the argument \code{abiotic_data}, these are also joined to the
neighborhood metric and growth data. The result is a design matrix that is
given to the \code{growth_mort_model} function. This entire process is
repeated for each neighborhood size (\code{radii}) and \code{focal_sps}, with
the mean square error of each resulting model being recorded. The function
outputs a list containing a data frame of all mean square error values and
a plot of mean square error vs. neighborhood size for each \code{focal_sps}.
}
\examples{
# See vignette "Selecting neighborhood size"
}
