% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_by_species.R
\name{site_by_species}
\alias{site_by_species}
\title{Create site x species matrix}
\usage{
site_by_species(neighbors, id_column, abundance = F)
}
\arguments{
\item{neighbors}{A neighborhoods object output by the \code{neighborhoods}
function}

\item{id_column}{Name of column in \code{neighbors} containing site names as
a string}

\item{abundance}{Boolean specifying whether an abundance or presence/absence
site x species matrix is desired (default is presence/absence)}
}
\value{
A site x species matrix where each row represents a distinct site or
neighborhood and each column represents a tree species. Row names are the 
site names. If \code{abundance = F} presence is indicated by 1 and absence 
by 0. If \code{abundance = T} values represent the number of trees of that
species in the neighborhood (excluding the focal if neighborhoods are
centered on a focal tree).
}
\description{
Takes in a neighborhoods object output by the \code{neighborhoods} function
and returns a site x species matrix where each site is a neighborhood. The 
user can specify whether the values in the matrix represent abundance or
presence/absence.
}
