% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_dist.R
\name{size_dist}
\alias{size_dist}
\title{Plot size class distributions}
\usage{
size_dist(size_data, stands, bin_size = 10)
}
\arguments{
\item{size_data}{Data frame containing size information of trees. Must 
include the columns: \code{tree_id}, \code{stand_id}, \code{dbh}, and
\code{year} (i.e. year in which size measurement was taken).}

\item{stands}{Character vector of names of stands for which size class
distributions are desired. To maintain plot clarity it is recommended that
10 stands or fewer are specified.}

\item{bin_size}{Integer representing desired width of size classes. Default
is 10 dbh units.}
}
\value{
A single plot containing a panel for each of the specified
\code{stands}.
}
\description{
Creates tree size class distribution plots for one or more stands.
}
\details{
The user-provided data frame \code{size_data} can contain trees from an 
unlimited number of stands but it is recommended that no more that 10 stands
be specified in the \code{stands} argument to maintain resolution in the 
resulting multi-panel plot. If \code{size_data} contains multiple size
measurements for a single tree, the \code{tree_id} and \code{year} columns
will be used to select the most recent size measurement for plotting.
}
\examples{
# Single stand
size_dist(size_data = tree, stands = "AB08", bin_size = 5)

# Multiple stands
size_dist(size_data = tree, stands = c("AM16", "AO03", "AR07"))
}
