% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_map.R
\name{stand_map}
\alias{stand_map}
\title{Create a stand map.}
\usage{
stand_map(map_data, dbh_data, x_limit, y_limit, labels = NULL)
}
\arguments{
\item{map_data}{Data frame of tree coordinates. Must include the columns:
\code{tree_id}, \code{x_coord}, and \code{y_coord}. If points (trees) are to
be labeled in the plot, \code{map_data} must also contain a column containing
the labels (e.g. tag numbers).}

\item{dbh_data}{Data frame of tree measurement data. Must include the
columns: \code{tree_id}, \code{dbh}, and \code{year}.}

\item{x_limit}{Numeric vector of length two indicating the x coordinate
range to be included.}

\item{y_limit}{Numeric vector of length two indicating the y coordinate
range to be included.}

\item{labels}{Name of column in \code{map_data} containing labels for trees
in the stem map, provided as a string. If \code{labels} argument is not 
provided, trees will not be labeled in the stem map.}
}
\value{
Stand map will be plotted in the plotting window.
}
\description{
Creates a map of the forest stand where each tree is a point, sized relative
to its most recent diameter at breast height (dbh) measurement. Trees for
which dbh data is missing from \code{dbh_data} will be plotted in blue
instead of the standard red and assigned a size equal to mean dbh in the
plot. Points can optinally be labeled using a column included in
\code{map_data}, such as tag numbers of the trees.
}
\details{
The x and y limits of the stand map can be specified because if a large
stand (>50m in width or length) is mapped all at once the labels probably
will not be readable.
}
\examples{
library(dplyr)
# Isolate mapping and tree data for one stand
one_stand_map <- mapping \%>\%
  filter(stand_id == "AB08")
one_stand_tree <- tree \%>\%
  filter(stand_id == "AB08")

# Create map of one quarter of the stand using tag numbers as labels
stand_map(one_stand_map, one_stand_tree, c(50, 100), c(50, 100), "tag")

}
