% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{data}
\name{tree}
\alias{tree}
\title{Repeated size measurements and mortality records for 9155 trees.}
\format{
A data frame with 60114 rows and 7 variables:
\describe{
  \item{tree_id}{unique identification code of the tree}
  \item{stand_id}{name of stand in which tree is located}
  \item{tag}{unique tag number of the tree}
  \item{species}{species identity of tree as a four letter code}
  \item{year}{year when measurement was taken}
  \item{dbh}{diameter at breast height of the tree at the time of
  measurement, in cm}
  \item{mort}{tree mortality status with 1 meaning dead and 0 meaning alive}
}
}
\source{
Data are available from the Pacific Northwest Permanent Sample Plot
Network on request \url{http://pnwpsp.forestry.oregonstate.edu/data}
}
\usage{
tree
}
\description{
A dataset containing repeated size measurements and mortality records for
9155 trees in 15 mapped forest stands located in Mount Rainier National Park,
WA, USA. These stands are part of the Pacific Northwest Permanent Sample Plot
Network (\url{http://pnwpsp.forestry.oregonstate.edu/}). Each row represents
a single measurement of a single tree and a determination of whether it was
alive or dead. PLEASE NOTE: the mortality data in the \code{mort} column were
randomly generated for the purposes of demonstrating functions in this
package and should not be used for real analysis. Although mortality data
are collected in these stands, mortality rates are extremely low and
therefore the real mortality data could not be used for demonstrating the
mortality model function in this package
}
\keyword{datasets}
