% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_check.R
\name{tree_check}
\alias{tree_check}
\title{Check tree measurement data}
\usage{
tree_check(tree_data, map_data)
}
\arguments{
\item{tree_data}{Data frame containing tree measurement data where each row
represents a single measurement of a single tree. Should contain the columns
\code{tree_id}, \code{stand_id}, \code{species}, \code{year} and \code{dbh}.
Any additional columns will be ignored by this function.}

\item{map_data}{Data frame containing tree mapping data. Should contain the
columns \code{tree_id}, \code{stand_id}, \code{species}, \code{x_coord},
and \code{y_coord}. Any additional columns will be ignored by this function.
Note that this function does not check the mapping data, which should first 
be checked with the function \code{mapping_check}.}
}
\value{
A list containing two elements:
\itemize{
  \item{\code{problem_trees} is a data frame containing the tree ids found
  to have data issues and a description of the issue}
  \item{\code{issue_summary} is a data frame that shows the number and 
  percentage of trees with at least one issue and with each of the specific
  issues}
}
}
\description{
Checks a tree measurements dataset for missing data and formatting that could 
lead to errors when applying other functions in this package. Outputs a
table of the trees with data issues and a table summarizing the number of
trees in the mapping dataset that have data issues.
}
\details{
The data issues checked for are: presence of required columns, single tree
ids referring to multiple trees, trees having no associated mapping data,
missing dbh, stand id, species, or measurement year. This function does
not check for misspelled stand ids or species, which should be checked 
independently. The common issue of negative growth rates resulting from
measurement error are not checked here, but are checked by
\code{growth_summary}.

Tree ids indicated to have data issues according to this function are not 
necessarily unusable. For instance, missing year data could be inferred from
knowledge on when certain stands were measured. The tree ids with data issues
should therefore be investigated further rather than being excluded from 
further analyses right away.
}
\examples{
tree_check_test <- tree_check(messy_tree, mapping)
}
