% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utm.R
\name{tree_utm}
\alias{tree_utm}
\title{Get UTM coordinates for individual trees}
\usage{
tree_utm(tree_x_y, stand_locs, original_crs, utm_crs)
}
\arguments{
\item{tree_x_y}{Data frame containing mapping information for trees. Must
contain the columns \code{stand_id} (name of stand the tree is located in),
\code{x_coord} (local x coordinate), and \code{y_coord} (local y coordinate).
Any additional columns will appear unchanged in the output.}

\item{stand_locs}{Data frame containing locations of stand origins (i.e. 
local coordinates (0, 0)) and their y-azimuths (i.e. bearing from (0, 0) to
(0, maximum y)). Columns must be named \code{stand_id}, \code{y_azim}, 
\code{latitude}, \code{longitude} - see dataset \code{stand_locations} for
an example. Latitudes and longitudes must be in decimal degrees or UTM units.}

\item{original_crs}{Coordinate reference system of the location data in 
\code{stand_locs} as an ESPG code.}

\item{utm_crs}{ESPG code of the UTM zone in which the trees are located. If
the location data are in UTM, this will be the same as \code{original_crs}.}
}
\value{
The input data frame \code{tree_x_y} with one additional column 
containing spatial information.
}
\description{
Calculates UTM coordinates for all trees in a user-provided mapping dataset
and returns a spatial object that can be plotted and joined with other
spatial datasets such as fine-scale topography data.
}
\details{
Location data for stands (argument \code{stand_locs}) must be provided in
either decimal degrees or UTM format. If your data are in another format 
such as degrees/minutes/seconds, please convert them to one of the accepted
formats. The \code{measurements::conv_unit} function is particularly
useful for achieving this.
}
\examples{
mapping_utm <- tree_utm(mapping, stand_locations, 4326, 32610)
}
