% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_model.R
\name{growth_model}
\alias{growth_model}
\title{Model tree growth}
\usage{
growth_model(training, outcome_var, focal_sps, rare_comps = "none")
}
\arguments{
\item{training}{A dataframe containing a separate row for each focal x
neighbor tree pair and a column for each variable to include in the
model, including the outcome variable. Outcome variable must be numeric,
other variables can be of any type, including character. Character variables
will be converted to factors before fitting the model.}

\item{outcome_var}{Name of column in \code{training} containing the outcome
variable, provided as a string.}

\item{focal_sps}{Name of species for which growth should be modeled,
provided as a string.}

\item{rare_comps}{Minimum number of interactions a competitor species must
appear in to remain separate from the "OTHR" category (see details). If not
specified, no "OTHR" category will be created and all competitor species will
remain separate.}
}
\value{
A list containing four elements: \code{mod} is the fitted glmnet
object, \code{obs_pred} is a dataframe containing observed and predicted 
growth, \code{R_squared} is the coefficient of determination, \code{mod_coef}
is a matrix containing the explanatory variables and their fitted
coefficients. These elements can be called with e.g. \code{$R_squared}.
}
\description{
Fits a regularized linear regression model of tree growth for a single
species, returning the model object, growth predictions, the coefficient
of determination, and the fitted model coefficients.
}
\details{
All variables other than the indicated outcome variable are included as
explanatory variables in the model. Regularized regression model is fitted
using the glmnet package. Predictions, R-squared, and model coefficients
returned are all based on the \code{"lambda.1se"} model but all models with
different lambda values are included in the returned mod object - see glmnet 
documentation for details.

Rare competitor species can be grouped together as "OTHR" for modeling if
desired. The optional argument \code{rare_comps} is a number indicating the
minimum number of interactions a species must appear in to remain separate
from the "OTHR" category. If species-specific densities are included in
\code{training} those of rare competitors will be combined into a new 
"OTHR_density" variable.
}
