% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_summary.R
\name{growth_summary}
\alias{growth_summary}
\title{Calculate annual growth}
\usage{
growth_summary(data)
}
\arguments{
\item{data}{A dataframe containing repeated measurements of growth.}
}
\value{
A dataframe containing growth rate measurements for each tree
}
\description{
Function requires a data frame where separate measurements of the same tree
appear in different rows. The data frame needs to have a column called "treeid"
that contains unique tree ID values, a column named "year" containing the year
of the measurement, and a column named "dbh" containing the dbh measurements.
}
