% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{neighborhoods}
\alias{neighborhoods}
\title{Function for creating neighborhoods}
\usage{
neighborhoods(mapping, stands = "all", radius, densities = F)
}
\description{
Option to specify stands (as a single stand ID, vector of IDs, or simply the
default of all stands in the mapping data), neighborhood radius, and whether
tree densities (all trees and species-specific densities) should be included
}
