% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{neighborhoods}
\alias{neighborhoods}
\title{Create neighborhoods}
\usage{
neighborhoods(mapping, stands = "all", radius, coords = NULL, tree = NULL)
}
\arguments{
\item{mapping}{Data frame containing tree coordinates.}

\item{stands}{Vector of names of stands for which neighborhoods are desired.}

\item{radius}{Numeric vector describing neighborhood radius in meters.}

\item{coords}{Data frame containing coordinates for which neighborhoods are
desired. Data frame should contain three columns in the following order -
location ids, x-coordinates, y-coordinates. Column names are unimportant.}

\item{tree}{Tree census data frame containing dbh measurements for each
tree id in \code{mapping}.}
}
\value{
Neighborhood information for all focal trees in \code{mapping} or,
if \code{coords} is provided, for all locations defined by coordinates.
}
\description{
Takes a data frame containing mapping information of trees in one or more
stands and returns neighborhood information for all trees or a set of 
user-provided coordinates.
}
\details{
This function returns a neighborhoods object, which is a data frame where
each focal tree or user-provided set of coordinates appears on multiple 
lines with each line containing information on one of the trees in its
neighborhood. An important piece of information on the neighboring trees is
their size (dbh). If the provided \code{mapping} dataset contains a dbh
column, and the \code{tree} argument is empty, the sizes of neighbors will be
taken from \code{mapping}. However, dbh data in mapping datasets usually
refers to the size at first measurement, which may not be ideal when
quantifying neighborhoods. If a tree census dataset is provided under the
optional argument \code{tree}, the most recent dbh measurement of each tree
in this dataset will be used for the sizes of neighbors.

The output neighborhoods object can be passed into a number of other
functions in ForestPlot, including \code{neighborhood_summary} and 
\code{site_by_species}.
}
\examples{
# Create neighborhoods for trees in mapping
neighborhoods(mapping, stands = c("AB08", "PP17"), radius = 10, tree = tree)

# Create neighborhoods for user-provided coordinates
locations <- data.frame(
loc_id = paste("A", 1:81, sep = ""),
x_coord = rep(seq(10, 90, 10), times = 9),
y_coord = rep(seq(10, 90, 10), each = 9))
neighborhoods(mapping, stands = "AB08", radius = 10, coords = locations)
}
