% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{neighborhoods}
\alias{neighborhoods}
\title{Create neighborhoods}
\usage{
neighborhoods(mapping, stands = "all", radius)
}
\arguments{
\item{mapping}{Data frame containing tree coordinates.}

\item{stands}{Vector of names of stands for which neighborhoods are desired.}

\item{radius}{Numeric vector describing neighborhood radius in meters.}
}
\value{
Neighborhood information for all focal trees in \code{mapping}.
}
\description{
Takes a data frame containing mapping information of trees in one or more
stands and returns a data frame with a row for each focal-neighbor pair
}
