% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{neighborhoods}
\alias{neighborhoods}
\title{Create neighborhoods}
\usage{
neighborhoods(mapping, stands = "all", radius, coords = NULL)
}
\arguments{
\item{mapping}{Data frame containing tree coordinates.}

\item{stands}{Vector of names of stands for which neighborhoods are desired.}

\item{radius}{Numeric vector describing neighborhood radius in meters.}

\item{coords}{Data frame containing coordinates for which neighborhoods are
desired. Data frame should contain three columns in the following order -
location ids, x-coordinates, y-coordinates. Column names are unimportant.}
}
\value{
Neighborhood information for all focal trees in \code{mapping} or,
if \code{coords} is provided, for all locations defined by coordinates.
}
\description{
Takes a data frame containing mapping information of trees in one or more
stands and returns a data frame with a row for each focal-neighbor pair
}
\examples{
# Create neighborhoods for trees in mapping
neighborhoods(mapping, stands = c("AB08", "PP17"), radius = 10)

# Create neighborhoods for user-provided coordinates
locations <- data.frame(
loc_id = paste("A", 1:81, sep = ""),
x_coord = rep(seq(10, 90, 10), times = 9),
y_coord = rep(seq(10, 90, 10), each = 9))
neighborhoods(mapping, stands = "AB08", radius = 10, coords = locations)
}
