% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_map.R
\name{stand_map}
\alias{stand_map}
\title{Create a stand map.}
\usage{
stand_map(map_data, x_limit, y_limit)
}
\arguments{
\item{map_data}{Data frame of tree coordinates. Must include the columns:
\code{x_coord}, \code{y_coord}, \code{dbh}, and \code{tag} (tag numbers will
be used to label trees on stand map).}

\item{x_limit}{Numeric vector of length two indicating the x coordinate
range to be included.}

\item{y_limit}{Numeric vector of length two indicating the y coordinate
range to be included.}
}
\value{
Stand map will be plotted in the plotting window.
}
\description{
Creates a map of the forest stand where each tree is a point, sized relative
to its diameter at breast height (dbh) and labeled with its tag number.
}
\details{
The x and y limits of the stand map can be specified because if a large
stand (>50m in width or length) is mapped all at once the labels probably
will not be readable.
}
\examples{
# Isolate mapping data for one stand
one_stand <- mapping \%>\%
  filter(stand_id == "AB08")

# Create stand map
stand_map(one_stand, c(50, 100), c(50, 100))
}
